hunt profile data all done

+++ utilities/save_to_excel.py
@@ -12,15 +12,15 @@ def save_profile_data_to_excel(data, file_path):
     if not os.path.exists(file_path):
         # File doesn't exist, create it with headers
         new_row.to_excel(file_path, index=False)
-        print(f"Created new file and saved data for: {data['Username']}")
+        print(f"✅ Created new file and saved data for: {data['Username']}")
     else:
         # File exists, append data
         existing_df = pd.read_excel(file_path)
         
-        # if data["Username"] in existing_df["Username"].values:
-        #     print(f"Username {data['Username']} already exists in file. Skipping save.")
-        #     return
+        if data["Username"] in existing_df["Username"].values:
+            print(f"❌ Username {data['Username']} already exists in file. Skipping save.")
+            return
 
         updated_df = pd.concat([existing_df, new_row], ignore_index=True)
         updated_df.to_excel(file_path, index=False)
-        print(f"Appended data for: {data['Username']}")
+        print(f"✅ Appended data for: {data['Username']}")
